+++ packages/stream_chat_flutter/example/lib/tutorial_part_6.dart
@@ -63,6 +63,12 @@ class MyApp extends StatelessWidget {
           borderRadius: BorderRadius.circular(8),
         ),
       ),
+      messageListViewTheme: const MessageListViewThemeData(
+          backgroundColor: Colors.grey,
+          backgroundImage: DecorationImage(
+            image: AssetImage('assets/background_doodle.png'),
+            fit: BoxFit.cover,
+          )),
       otherMessageTheme: MessageThemeData(
         messageBackgroundColor: colorTheme.textHighEmphasis,
         messageTextStyle: TextStyle(