+++ packages/stream_chat_flutter/lib/src/user_avatar.dart
@@ -92,7 +92,9 @@ class UserAvatar extends StatelessWidget {
                         streamChatTheme
                             .ownMessageTheme.avatarTheme?.borderRadius,
                     image: DecorationImage(
-                        image: imageProvider, fit: BoxFit.cover),
+                      image: imageProvider,
+                      fit: BoxFit.cover,
+                    ),
                   ),
                 ),
               )