+++ packages/stream_chat_persistence/lib/src/stream_chat_persistence_client.dart
@@ -229,7 +229,7 @@ class StreamChatPersistenceClient extends ChatPersistenceClient {
         final parentId = message.parentId!;
         messageByParentIdDictionary[parentId] = [
           ...messageByParentIdDictionary[parentId] ?? [],
-          message
+          message,
         ];
       }
       return messageByParentIdDictionary;